Initial patch set to build libkmod and kmod utilities on MacOS X
This is a port of kmod for Linux to MacOS X.  Obviously loading and
unloading kernel modules will not work.  The objective is the libkmod
functions to work on built Linux kernel modules.  Load a .ko file,
read it, and so on.  This is intended to support developers wanting to
work with Linux kernel modules while on a MacOS X system.  It serves
no other practical purpose.